pyproject.toml
[tool.poetry]
name = "k8smap"
version = "0.1.1"
description = ""
authors = ["Paul Beck"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
mypy = "^0.991"
black = "^22.12.0"
click = "^8.1.3"
flake8 = "^6.0.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
k8smap = "k8smap.main:cli"